refact: Refact the NinjaController to include the ResponseEntity to all endpoints and set the StatusCode for each request

+++ src/main/java/dev/java10x/CadastroDeNinjas/Ninjas/NinjaController.java
@@ -1,5 +1,7 @@
 package dev.java10x.CadastroDeNinjas.Ninjas;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -22,33 +24,57 @@ public class NinjaController {
 
     //Cadastrar Ninjas
     @PostMapping("/criar")
-    public NinjaDTO createNinja(@RequestBody NinjaDTO ninjaDTO) {
-        return ninjaService.createNinja(ninjaDTO);
+    public ResponseEntity<String> createNinja(@RequestBody NinjaDTO ninjaDTO) {
+        NinjaDTO novoNinja = ninjaService.createNinja(ninjaDTO);
+        return ResponseEntity.status(HttpStatus.CREATED)
+            .body("Ninja criado com sucesso: " + novoNinja.getNome() + " (ID) " + novoNinja.getId());
     }
 
     //Mostrar todos os ninjas
     @GetMapping("/listar")
-    public List<NinjaDTO> listNinjas() {
-        return ninjaService.listNinjas();
+    public ResponseEntity<List<NinjaDTO>> listNinjas() {
+        List<NinjaDTO> ninjas = ninjaService.listNinjas();
+        return ResponseEntity.ok(ninjas);
     }
 
     //Monstrar Ninja por ID
     @GetMapping("/listar/{id}")
-    public NinjaDTO listNinjaByID(@PathVariable Long id) {
-        return ninjaService.listNinjaById(id);
+    public ResponseEntity<?> listNinjaByID(@PathVariable Long id) {
+        NinjaDTO ninja = ninjaService.listNinjaById(id);
+        if (ninja != null) {
+            ninjaService.listNinjaById(id);
+            return ResponseEntity.ok(ninja);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                .body("Ninja com o ID: " + id + " Não foi encontrado!");
+        }
     }
 
     //Alterar dados do ninja
     @PutMapping("/alterar/{id}")
-    public NinjaDTO updateNinja(@PathVariable Long id, @RequestBody NinjaDTO ninjaDTO) {
-        return ninjaService.updateNinja(id, ninjaDTO);
+    public ResponseEntity<?> updateNinja(@PathVariable Long id, @RequestBody NinjaDTO ninjaDTO) {
+        NinjaDTO ninja = ninjaService.listNinjaById(id);
+        if (ninja != null) {
+            ninjaService.updateNinja(id, ninjaDTO);
+            return ResponseEntity.ok(ninja);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                .body("Ninja com o ID: " + id + " Não foi encontrado!");
+        }
 
     }
 
     //Deletar Ninja
     @DeleteMapping("/deletar/{id}")
-    public void deleteNinjaById(@PathVariable Long id) {
-        ninjaService.deleteNinjaById(id);
+    public ResponseEntity<String> deleteNinjaById(@PathVariable Long id) {
+        NinjaDTO ninja = ninjaService.listNinjaById(id);
+        if(ninja != null) {
+            ninjaService.deleteNinjaById(id);
+            return ResponseEntity.ok("Ninja com o ID: " + id + " Foi deletado com sucesso!");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                .body("Ninja com o ID: " + id + " Não foi encontrado!");
+        }
     }
 
 }
